Update index.html
Major updates to card visuals; textBoxes are now fit-content height (with min- and max- depending on card type). Because of this, card filigrees and backgrounds have been changed.
The full card has an image background and a gradient background, and the textBox has the filigree, inset (if applicable), and transparent/white gradient to fill the background. Medallions have also been moved inside the textBox so they move with the filigree.